+++ src/templates/gitignore/GitignoreTemplate.ts
@@ -4,11 +4,21 @@ import { ITemplate } from "../../interfaces/templates/ITemplate";
 import { Template } from "../Template";
 
 export class GitignoreTemplate extends Template implements ITemplate {
-  public render(templateData: { publish_to_git: boolean; list_for_load_on_demand: string }): void {
+  public async render(templateData: {
+    publish_to_git: boolean;
+    list_for_load_on_demand: string;
+    force: boolean;
+  }): Promise<void> {
     if (fs.existsSync(this.CWD)) {
       const renderedContent: string = ejs.render(this.sourceContent, templateData);
 
-      fs.writeFileSync(this.destinationFilePath, renderedContent, { encoding: "utf-8" });
+      if (this.isExistDestFile()) {
+        if (templateData.force) {
+          fs.writeFileSync(this.destinationFilePath, renderedContent, { encoding: "utf-8" });
+        }
+      } else {
+        fs.writeFileSync(this.destinationFilePath, renderedContent, { encoding: "utf-8" });
+      }
     }
   }
 }
